Statistics without proper frontend fix

+++ FootBookEJBProject/ejbModule/facade/Facade.java
@@ -44,11 +44,11 @@ public class Facade implements FacadeLocal {
 		return pitchEAO.getAllPitches();
 	};
 	
-	public int getUserCount() {
+	public long getUserCount() {
 		return userEAO.getUserCount();
 	}
 	
-	public int getUsersOnMatchesCount() {
+	public long getUsersOnMatchesCount() {
 		return userEAO.getUsersOnMatchesCount();
 	}
