+++ FootBookEJBProject/ejbModule/ics/eao/UserEAOLocal.java
@@ -10,8 +10,8 @@ public interface UserEAOLocal {
 	public void updateUser(User user);
 	public void deleteUser(String userId);
 	public User findUserById(String userId);
-	public int getUserCount(); // For statistics of how many users are registered
-	public int getUsersOnMatchesCount(); //Stats on how many users are registered on matches
+	public long getUserCount(); // For statistics of how many users are registered
+	public long getUsersOnMatchesCount(); //Stats on how many users are registered on matches
